src/apps/nfc/action/resolvers/create-action.resolver.spec.ts
import { Test, TestingModule } from '@nestjs/testing';

// custom items
import { CreateActionResolver } from './create-action.resolver';
import { ICommandBus } from '@hades/shared/domain/bus/command-bus';
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { actions } from '@hades/nfc/action/infrastructure/seeds/action.seed'
import { NfcCreateActionInput } from './../../../../../src/graphql';

describe('CreateActionResolver', () => 
{
    let resolver: CreateActionResolver;
    let queryBus: IQueryBus;
    let commandBus: ICommandBus;

    beforeEach(async () => 
    {
        const module: TestingModule = await Test.createTestingModule({
            providers: [
                CreateActionResolver,
                {
                    provide: IQueryBus,
                    useValue: {
                        ask: () => {},
                    }
                },
                {
                    provide: ICommandBus,
                    useValue: {
                        dispatch: () => {},
                    }
                },
            ]
        }).compile();

        resolver    = module.get<CreateActionResolver>(CreateActionResolver);
        queryBus    = module.get<IQueryBus>(IQueryBus);
        commandBus  = module.get<ICommandBus>(ICommandBus);
    });

    it('CreateActionResolver should be defined', () => 
    {
        expect(resolver).toBeDefined();
    });

    // Test get method
    describe('main', () => 
    {
        it('CreateActionResolver should be defined', () => 
        {
            expect(resolver).toBeDefined();
        });

        it('should return an action created', async () => 
        {
            jest.spyOn(queryBus, 'ask').mockImplementation(() => new Promise(resolve => resolve(actions[0])));
            expect(await resolver.main(<NfcCreateActionInput>actions[0])).toBe(actions[0]);
        });
    });
});